requirements.txt
anyascii==0.3.3
asgiref==3.9.1
beautifulsoup4==4.13.4
black==25.12.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.3.1
cloudinary==1.44.1
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2.4
django-cloudinary-storage==0.3.0
django-filter==25.1
django-modelcluster==6.4
django-permissionedforms==0.1
django-stubs-ext==5.2.2
django-taggit==6.1.0
django-tasks==0.7.0
django-treebeard==4.7.1
djangorestframework==3.16.0
draftjs_exporter==5.1.0
et_xmlfile==2.0.0
filetype==1.2.0
flake8==7.3.0
gunicorn==23.0.0
idna==3.10
isort==7.0.0
laces==0.1.2
mccabe==0.7.0
mypy_extensions==1.1.0
openpyxl==3.1.5
packaging==25.0
pathspec==1.0.3
pillow==11.3.0
platformdirs==4.5.1
psycopg2-binary==2.9.10
pycodestyle==2.14.0
pyflakes==3.4.0
pytokens==0.3.0
requests==2.32.4
segno==1.6.6
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
telepath==0.3.1
typing_extensions==4.14.1
urllib3==2.5.0
wagtail==7.0.2
whitenoise==6.6.0
Willow==1.11.0
